+++ libraries/printHelpers/CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.5.1] - 2025-12-29
+- fix #26, add units() e.g. to replace E+xx => mVolt 
+- fix #27, add rightAlign to eng() notation.
+- update examples
+- update GitHub actions
+- minor edits
+
+
 ## [0.5.0] - 2025-06-30
 - fix #24, add printHelpersMT.h,
   - class implementation for multThreading.